+++ src/components/Events/Event.jsx
@@ -3,46 +3,13 @@ import moment from 'moment';
 import { faCalendarAlt, faUser } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import './Event.css';
-
 import OnlineLogo from './OnlineLogo';
 
-const TEXT_LENGTH = 48;
+import eventTypes from '../../eventTypes';
 
-const EVENT_TYPES = {
-  1: {
-    display: 'Sosialt',
-    color: 'rgb(67, 177, 113)',
-  },
-  2: {
-    display: 'Bedriftspresentasjon',
-    color: 'rgb(235, 83, 110)'
-  },
-  3: {
-    display: 'Kurs',
-    color: 'rgb(18, 125, 189)'
-  },
-  4: {
-    display: 'Utflukt',
-    color: 'rgb(253, 189, 71)'
-  },
-  5: {
-    display: 'Ekskursjon',
-    color: 'rgb(42, 198, 249)'
-  },
-  6: {
-    display: 'Internt',
-    color: 'rgb(231, 94, 59)'
-  },
-  7: {
-    display: 'Annet',
-    color: 'rgb(179, 107, 205)'
-  },
-  8: {
-    display: 'Realfagskjelleren',
-    color: 'rgb(231, 94, 59)'
-  },
-};
+import './Event.css';
+
+const TEXT_LENGTH = 48;
 
 function Event({event}) {
   const formattedDate = moment(event.start_date).format('DD.MM');
@@ -55,7 +22,7 @@ function Event({event}) {
     <div className="event">
       {imageSrc ?
         <img className="thumb" src={imageSrc} alt="Event" /> :
-        <OnlineLogo className="thumb" fillColor={EVENT_TYPES[event.event_type]?.color} />
+        <OnlineLogo className="thumb" fillColor={eventTypes[event.event_type]?.color} />
       }
       <div className="data">
         <h3 className="title">{event.title}</h3>
