+++ src/eventTypes.js
@@ -1,0 +1,36 @@
+const EVENT_TYPES = {
+  1: {
+    display: 'Sosialt',
+    color: 'rgb(67, 177, 113)',
+  },
+  2: {
+    display: 'Bedriftspresentasjon',
+    color: 'rgb(235, 83, 110)'
+  },
+  3: {
+    display: 'Kurs',
+    color: 'rgb(18, 125, 189)'
+  },
+  4: {
+    display: 'Utflukt',
+    color: 'rgb(253, 189, 71)'
+  },
+  5: {
+    display: 'Ekskursjon',
+    color: 'rgb(42, 198, 249)'
+  },
+  6: {
+    display: 'Internt',
+    color: 'rgb(231, 94, 59)'
+  },
+  7: {
+    display: 'Annet',
+    color: 'rgb(179, 107, 205)'
+  },
+  8: {
+    display: 'Realfagskjelleren',
+    color: 'rgb(231, 94, 59)'
+  },
+};
+
+export default EVENT_TYPES;
